trace: Fix parameter types in hw/scsi
An upcoming patch will let the compiler warn us when we are silently
losing precision in traces; update the trace definitions to pass
through the full value at the callsite.  Also update some callers
to avoid variable-sized dma_addr_t (which cannot be easily used in
trace headers) and needless casts for values used in tracing.

Message-Id: <20170313195547.21466-26-eblake@redhat.com>

+++ hw/scsi/esp-pci.c
@@ -206,7 +206,7 @@ static void esp_pci_io_write(void *opaque, hwaddr addr,
         trace_esp_pci_sbac_write(pci->sbac, val);
         pci->sbac = val;
     } else {
-        trace_esp_pci_error_invalid_write((int)addr);
+        trace_esp_pci_error_invalid_write(addr);
     }
 }
